requirements.txt
alembic==0.6.3
SQLAlchemy==0.9.3
psycopg2==2.5.2
Flask==0.10.1
Flask-Assets==0.9
cssmin==0.2.0
jsmin==2.0.9
PyYAML==3.10
